update everything related to home.jsx with breaking it down into smaller components
